node: silence -Wsign-compare in node::get_child bounds check
Both get_child overloads compared `int index >= children.size()`
(size_t), a signed/unsigned comparison that warns under -Wsign-compare.
The preceding `index < 0` short-circuit already guarantees index is
non-negative at that point, so static_cast<size_t>(index) is
value-preserving — the cast makes the comparison size_t-vs-size_t and
removes the warning without changing behavior.

Closes #372.

+++ include/genogrove/structure/grove/node.hpp
@@ -340,7 +340,7 @@ class node {
      * @throws std::out_of_range if index is invalid
      */
     node* get_child(int index) {
-        if(index < 0 || index >= this->children.size()) {
+        if(index < 0 || static_cast<size_t>(index) >= this->children.size()) {
             throw std::out_of_range("child index out of range");
         }
         return this->children[index];
@@ -353,7 +353,7 @@ class node {
      * @throws std::out_of_range if index is invalid
      */
     node* get_child(int index) const {
-        if(index < 0 || index >= this->children.size()) {
+        if(index < 0 || static_cast<size_t>(index) >= this->children.size()) {
             throw std::out_of_range("child index out of range");
         }
         return this->children[index];
